feat(keys): Explicitly track timestamp of last key rotation.
Up until now we have been using `verifierSetAt` for the timestamp
component of OAuth scoped-key key ids. This works but means that we
generate a new kid whenever the user changes their password, even
if that did not involve resetting the master encryption key `kB`.

With this change, we track a separate timestamp that is only bumped
when kB actually changes, which should produce more stable key ids
while preserving the ability to quickly compare them to determine
which is the most recent.

Fixes #490.